skill-evaluator/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "skill-evaluator"
version = "0.1.0"
description = "Evaluate agent skills before you install them. Think 'npm audit' for SKILL.md files."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Causal Data Science Skills Team"}
]
keywords = ["agent", "skills", "evaluation", "SKILL.md", "AI", "quality"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "pyyaml>=6.0",
    "rich>=13.0",
    "click>=8.0",
]

[project.optional-dependencies]
llm = [
    "openai>=1.0",
    "anthropic>=0.20",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
skill-eval = "skill_evaluator.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
